update 10/5 1:14pm
For the bodyDown and bodyUp methods, I managed to whittle the number of for loops present in each down to 5 instead of 6. I figured since it took up less space it would be more efficient? I'm not sure if that's sound logic, I would love to hear your thoughts on this.
Also, at the end of each of the aforementioned methods, as well as the divider method, there's a println at the end. I removed the print statement before it and moved its output of "|" and "+" into the println statement. Again, I thought this would be more efficient and save space but it could just be a needless addition, I'm unsure.